# verishda-slint/openapi.yaml
openapi: 3.0.3
info:
  title: Verishda
  description: |-
    API für das Verishda-Backend. 

  version: '0.1'
externalDocs:
  description: Find out more about Swagger
  url: https://github.com/werischda/werischda
servers:
  - url: https://verishda.fermyon.app/
  - url: https://verishda.shuttleapp.rs/
  - url: http://127.0.0.1:3000/
  - url: http://127.0.0.1:8000/
paths:
  /api/sites:
    get:
      summary: Get available sites and their geolocation
      operationId: handle_get_sites
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Site'          
      security:
        - petstore_auth:
            - write:pets
            - read:pets
  /api/sites/{siteId}/presence:
    get:
      summary: See who is present at the specified site
      operationId: handle_get_sites_siteid_presence
      parameters:
        - $ref: '#/components/parameters/SitePathParam'
      responses:
        '200':
          $ref: '#/components/responses/PresenceResponse'
        '404':
          description: Site not found
      security:
        - petstore_auth:
            - write:pets
            - read:pets
  /api/sites/{siteId}/hello:
    post:
      summary: 'Call-In: Tell site that user is there at the moment'
      operationId: handle_post_sites_siteid_hello
      parameters:
        - $ref: '#/components/parameters/SitePathParam'
      responses:
        '200':
          $ref: '#/components/responses/PresenceResponse'
        '404':
          description: Site not found
      security:
        - petstore_auth:
            - write:pets
            - read:pets
  /api/sites/{siteId}/announce:
    put:
      summary: 'Announce a future presence(s) for this sites'
      description: >-
        Announcing presence via this endpoint will replace all previous
        presence announcements for this site for the current user.
      operationId: handle_put_announce
      parameters:
        - $ref: '#/components/parameters/SitePathParam'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PresenceAnnouncements'
      responses:
        '204':
          description: presence announced successfully
        '400':
          description: Site not found
      security:
        - petstore_auth:
            - write:pets
            - read:pets
            
components:
  responses:
    PresenceResponse:
      description: Successful operation
      content:
        application/json:
          schema:
            type: array
            items:
              $ref: '#/components/schemas/Presence'          
  parameters:
    SitePathParam:
      name: siteId
      schema:
        type: string
      in: path
      required: true
  schemas:
    Site:
      required:
        - id
        - name
        - longitude
        - latitude
      type: object
      properties:
        id:
          type: string
          example: 'A236E288-913C-4CFF-AD68-33817B87C0D8'
        name:
          type: string
          example: 'Almato Reutlingen'
        longitude:
          type: number
          format: float
          example: 48.4883438
        latitude:
          type: number
          format: float
          example: 9.2146156
    Presence:
      description:
        Contains presence information for a particular person for the
        site this presence object belongs to. Note that a presence 
        object may be returned by the server, even if a person is
        not present (see currently_present field) if the user
        has a presence announced. Absent persons without presence
        announcements will not be delivered.
      required:
        - logged_as_name
        - is_self
        - currently_present
        - announced_dates
      properties:
        logged_as_name:
          type: string
          example: 'Maxi Mustermax'
        is_self:
          type: boolean
        currently_present:
          type: boolean
          description: >-
            Declares whether the named user is currently present. A
            user may not be present, but still have presence announcements.
        announced_dates:
          type: array
          items:
            type: string
            format: localdate
    PresenceAnnouncements:
      type: array
      description: >-
        A list of announcements, declaring when the current user plans
        to be present at a given site. Note that the 'date' field
        of the individual PresenceAnnouncement items may be duplicates
        for the same date; this way, users can announce that they
        plan to be present at multiple sites on the same day. The API
        is deliberately open to this, so e.g. team leads can announce
        plans to be on site A in the morning and to visit site B in the
        afternoon.
      items:
        $ref: '#/components/schemas/PresenceAnnouncement'
    PresenceAnnouncement:
      properties:
        date:
          description: Date for which the user announces their presence
          type: string
          format: localdate
      required:
      - date
  securitySchemes:
    petstore_auth:
      type: oauth2
      flows:
        implicit:
          authorizationUrl: https://lemur-5.cloud-iam.com/auth/realms/werischda/protocol/openid-connect/auth
          scopes:
            'write:pets': foo
            read:pets: bar